project/requirements.txt
asyncpg
asynctest
fastapi
requests
tortoise-orm[asyncpg]
uvicorn